+++ Utils/Rendering.h
@@ -10,6 +10,7 @@
 #include "../Core/Camera.h"
 #include "../Gameplay/Food.h"
 #include "../Gameplay/Snake.h"
+#include <cstddef>
 
 // 前向声明
 class Snake;
@@ -29,5 +30,5 @@ void DrawSnakeEyes(const Vector2& position, const Vector2& direction, float radi
 bool IsCircleInScreen(const Vector2& center, float r); // 检查圆是否在屏幕范围内
 void DrawUI(); // 绘制用户界面元素
 void DrawEnhancedFood(const Vector2& screenPos, float radius, int color, int index); // 绘制增强视觉效果的食物
-void DrawSnakeSegment(const Vector2& screenPos, float radius, int color, int segmentIndex); // 绘制蛇段
+void DrawSnakeSegment(const Vector2& screenPos, float radius, int color, size_t segmentIndex); // 绘制蛇段
 void DrawPauseMenu(); // 绘制暂停菜单
